+++ modules/core/src/main/java/org/terrier/querying/parser/QueryParser.java
@@ -77,6 +77,9 @@ public class QueryParser
 			String suffix = "'" + query + "'";
 			if (qid != null)
 				suffix = "qid " + qid + " " + suffix;
+			if (e instanceof TokenMgrError) {
+				suffix += " -- " + e.getMessage();
+			}
             throw new QueryParserException("Failed to process " + suffix,e);
 		}
         if (rtr == null)