+++ doc/whats_new.md
@@ -6,6 +6,8 @@ Terrier 5.6 - xx/xx/2021
 
 Under development.
 
+## Retrieval
+ - [#182](https://github.com/terrier-org/terrier-core/issues/182) TokenMgrError exception should be promoted to the QueryParserException (with thanks to Eugene Yang and Sean MacAvaney)
 
 Terrier 5.5 - 20/05/2021
 ------------------------